# src/main/resources/public/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Projekt Bus
  description: Application made as university project (Java Spring Rest MongoDB)
servers:
  - url: https://peaceful-sierra-14544.herokuapp.com/
paths:
...
  /bus-lines:
    get:
      tags:
        - bus-lines
      summary: |
        Get list of BusLines
      operationId: getBusLines
      responses:
        200:
          description: Returns BusLines list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusLine'

    post:
      tags:
        - bus-lines
      summary: Add new BusLine
      operationId: addBusLine
      requestBody:
        description: BusLine object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBusLine'
      responses:
        201:
          description: BusLine created
          headers:
            Location:
              schema:
                type: "string"
              description: URL adress of created BusLine
        400:
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "name: Pole wymagane",
                  "endBusStopId: Pole wymagane",
                  "startBusStopId: Pole wymagane"
                ]
              }
  /bus-lines/{name}:
    get:
      tags:
        - bus-lines
      summary: Get BusLine by name
      operationId: getBusLineByName
      parameters:
        - name: name
          in: path
          description: BusLine name
          required: true
          schema:
            type: string
      responses:
        200:
          description: BusLine response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusLine"
        404:
          description: BusLine not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znalezniono linii"
                ]
              }
    delete:
      tags:
        - bus-lines
      summary: Delete BusLine by name
      operationId: deleteBusLineByName
      parameters:
        - name: name
          in: path
          description: BusLine name
          required: true
          schema:
            type: string
      responses:
        200:
          description: BusLine deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
              example: {
                "response": "Linia została usunięta"
              }
        404:
          description: BusLine not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znalezniono linii"
                ]
              }
  /bus-lines/{busLineId}/bus-stops/{busStopId}:
    post:
      tags:
        - bus-lines
      summary: Add new BusStop to existing BusLine
      description: BusStopId must be different then BusStop used as BusLine start or end
      operationId: addBusStopToBusLine
      parameters:
        - name: busLineId
          in: path
          description: BusLine id
          required: true
          schema:
            type: string
        - name: busStopId
          in: path
          description: BusStop id
          required: true
          schema:
            type: string
      responses:
        200:
          description: BusStop added to BusLine
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
              example: {
                "response": "Przystanek dodano do linii"
              }
        404:
          description: BusLine or BusStop not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znalezniono linii"
                ]
              }
        409:
          description: BusStop already added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Przystanek juz należy do tej linii"
                ]
              }
    delete:
      tags:
        - bus-lines
      summary: Delete BusStop from BusLine by id
      operationId: deleteBusStopFromBusLineById
      parameters:
        - name: busLineId
          in: path
          description: BusLine id
          required: true
          schema:
            type: string
        - name: busStopId
          in: path
          description: BusStop id
          required: true
          schema:
            type: string
      responses:
        200:
          description: BusStop deleted from BusLine
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
              example: {
                "response": "Przystanek poprawnie usunięty z linii"
              }
        404:
          description: BusLine or BusStop not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znalezniono linii"
                ]
              }
        409:
          description: Bus stop not on list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
              example: {
                "errors": [
                  "Nie znalezniono linii"
                ]
              }
...
components:
  schemas:
...
    BusStop:
      properties:
        id:
          type: string
        name:
          type: string
        myGeoPoint:
          $ref: '#/components/schemas/MyGeoPoint'
...
    MyGeoPoint:
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: -90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
    BusLine:
      properties:
        id:
          type: string
        name:
          type: string
        startBusStopId:
          type: string
        endBusStopId:
          type: string
        busStopList:
          type: array
          items:
            $ref: '#/components/schemas/BusStop'
    NewBusLine:
      required:
        - name
        - startBusStopId
        - endBusStopId
      properties:
        name:
          type: string
        startBusStopId:
          type: string
        endBusStopId:
          type: string
...
    ApplicationResponse:
      properties:
        response:
          type: string
    ApplicationError:
      properties:
        errors:
          $ref: '#/components/schemas/ErrorList'
    ErrorList:
      type: array
      items:
        properties:
          field:
            type: string
...